RSSEC-158 fix: rename folder
