+++ src/pages/catalog/functions/buildProductPage.ts
